examples/requirements/flask.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile flask.in
#
attrs==22.1.0
    # via pytest
babel==2.11.0
    # via flask-babelex
click==8.1.3
    # via flask
coverage[toml]==6.5.0
    # via
    #   -r test.in
    #   pytest-cov
distlib==0.3.6
    # via virtualenv
factory-boy==3.2.1
    # via -r test.in
faker==15.3.4
    # via
    #   -r test.in
    #   factory-boy
filelock==3.8.0
    # via
    #   tox
    #   virtualenv
flask==2.2.2
    # via
    #   -r flask.in
    #   flask-admin
    #   flask-babelex
    #   flask-sqlalchemy
    #   flask-testing
flask-admin==1.6.0
    # via -r flask.in
flask-babelex==0.9.4
    # via -r flask.in
flask-sqlalchemy==3.0.2
    # via -r flask.in
flask-testing==0.8.1
    # via -r flask.in
greenlet==2.0.1
    # via sqlalchemy
iniconfig==1.1.1
    # via pytest
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via
    #   flask
    #   flask-babelex
lxml==4.9.1
    # via
    #   python-docx
    #   python-pptx
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
packaging==21.3
    # via
    #   pytest
    #   tox
parameterize==0.2
    # via -r test.in
pdfkit==1.0.0
    # via -r common.in
pillow==9.3.0
    # via python-pptx
platformdirs==2.5.4
    # via virtualenv
pluggy==1.0.0
    # via
    #   pytest
    #   tox
py==1.11.0
    # via
    #   -r test.in
    #   pytest
    #   tox
pyparsing==3.0.9
    # via packaging
pytest==6.2.5
    # via
    #   -r test.in
    #   pytest-cov
    #   pytest-django
    #   pytest-ordering
    #   pytest-pythonpath
pytest-cov==4.0.0
    # via -r test.in
pytest-django==4.5.2
    # via -r test.in
pytest-ordering==0.6
    # via -r test.in
pytest-pythonpath==0.7.4
    # via -r test.in
python-dateutil==2.8.2
    # via faker
python-docx==0.8.11
    # via -r common.in
python-pptx==0.6.21
    # via -r common.in
pytz==2022.6
    # via babel
six==1.16.0
    # via
    #   python-dateutil
    #   tox
speaklater==1.3
    # via flask-babelex
sqlalchemy==1.4.44
    # via
    #   flask-sqlalchemy
    #   sqlalchemy-utils
sqlalchemy-utils==0.38.3
    # via -r flask.in
toml==0.10.2
    # via pytest
tomli==2.0.1
    # via
    #   coverage
    #   tox
tox==3.27.1
    # via -r test.in
virtualenv==20.17.0
    # via tox
werkzeug==2.2.2
    # via flask
wtforms==3.0.1
    # via flask-admin
xlsxwriter==3.0.3
    # via python-pptx
